+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project are documented in this file, based on [Keep 
 
 
 ## [Unreleased]
+- Ignore directories created by `languageSpecBuilder.initialize`, fixing incrementality issues from clean builds.
 
 
 ## [0.5.3]